feat: 중간저장
