+++ src/app/meetings/[id]/api/meeting-detail.api.ts
@@ -6,7 +6,7 @@ import type {
   MeetingJoinResponse,
   MeetingListResponse,
   MeetingParticipantsResponse,
-} from "@/app/meetings/[meetingId]/types";
+} from "@/app/meetings/[id]/types";
 
 const PARTICIPANTS_PAGE_SIZE = 100;
 const RECOMMENDED_MEETINGS_PAGE_SIZE = 100;
