+++ src/app/meetings/[id]/components/MeetingHeaderSection.tsx
@@ -9,11 +9,11 @@ import heartsFalse from "@/assets/icon/hearts/hearts-false.svg";
 import heartsTrue from "@/assets/icon/hearts/hearts-true.svg";
 import meatballsLgIcon from "@/assets/icon/meatballs/meatballs-lg.svg";
 import profileFemaleSm from "@/assets/img/profile/female1-sm.jpg";
-import { EditMeetingModal } from "@/app/meetings/[meetingId]/components/EditMeetingModal";
+import { EditMeetingModal } from "@/app/meetings/[id]/components/EditMeetingModal";
 import {
   MeetingDetailData,
   MeetingParticipantUser,
-} from "@/app/meetings/[meetingId]/types";
+} from "@/app/meetings/[id]/types";
 import { BtnCommon } from "@/components/ui/BtnCommon";
 import {
   DropdownMenu,
