+++ src/components/skeleton/DetailCardSkeleton.tsx
