+++ src/lib/pagination.ts
@@ -1,0 +1,5 @@
+export const getNextPageParam = <
+  T extends { hasMore: boolean; nextCursor: string | null },
+>(
+  lastPage: T,
+) => (lastPage.hasMore ? (lastPage.nextCursor ?? undefined) : undefined);
